shift all java files into single "duke" package
